+++ src/components/List/ListItem.tsx
@@ -1,9 +1,7 @@
 import * as React from 'react'
 import * as PropTypes from 'prop-types'
-import * as cx from 'classnames'
-
-import { createShorthandFactory, customPropTypes, pxToRem, UIComponent } from '../../lib'
-import Layout from '../Layout'
+import { createShorthandFactory, customPropTypes, UIComponent } from '../../lib'
+import ItemLayout from '../ItemLayout'
 import { ListItemBehavior } from '../../lib/accessibility'
 import { Accessibility } from '../../lib/accessibility/interfaces'
 
@@ -35,9 +33,6 @@ class ListItem extends UIComponent<any, any> {
     /** A list item can appear more important and draw the user's attention. */
     important: PropTypes.bool,
     media: PropTypes.any,
-    renderContentArea: PropTypes.any,
-    renderHeaderArea: PropTypes.any,
-    renderMainArea: PropTypes.any,
 
     /** A list item can indicate that it can be selected. */
     selection: PropTypes.bool,
@@ -66,9 +61,6 @@ class ListItem extends UIComponent<any, any> {
     'headerMedia',
     'important',
     'media',
-    'renderContentArea',
-    'renderHeaderArea',
-    'renderMainArea',
     'selection',
     'styles',
     'truncateContent',
@@ -79,83 +71,6 @@ class ListItem extends UIComponent<any, any> {
   static defaultProps = {
     as: 'li',
     accessibility: ListItemBehavior as Accessibility,
-
-    renderMainArea: (props, state, classes) => {
-      const { renderHeaderArea, renderContentArea } = props
-
-      const headerArea = renderHeaderArea(props, state, classes)
-      const contentArea = renderContentArea(props, state, classes)
-
-      return (
-        <div
-          className="ui-list__item__main"
-          // vertical
-          // disappearing
-          // rootCSS={{
-          //   gridTemplateRows: "1fr 1fr"
-          // }}
-          style={{
-            gridTemplateRows: '1fr 1fr',
-          }}
-          // start={headerArea}
-          // end={contentArea}
-        >
-          {headerArea}
-          {contentArea}
-        </div>
-      )
-    },
-
-    renderHeaderArea: (props, state, classes) => {
-      const { debug, header, headerMedia, truncateHeader, selection } = props
-      const { isHovering } = state
-
-      const mergedClasses = cx('ui-list__item__header', classes.header)
-      const mediaClasses = cx('ui-list__item__headerMedia', classes.headerMedia)
-      const headerMediaStyle =
-        headerMedia && selection && isHovering ? { color: 'inherit' } : undefined
-
-      return !header && !headerMedia ? null : (
-        <Layout
-          className={mergedClasses}
-          alignItems="end"
-          gap={pxToRem(8)}
-          debug={debug}
-          // disappearing={!truncateHeader}
-          truncateMain={truncateHeader}
-          rootCSS={selection && isHovering ? { color: 'inherit' } : undefined}
-          main={header}
-          end={
-            headerMedia && (
-              <span style={headerMediaStyle} className={mediaClasses}>
-                {headerMedia}
-              </span>
-            )
-          }
-        />
-      )
-    },
-
-    renderContentArea: (props, state, classes) => {
-      const { debug, content, contentMedia, truncateContent, selection } = props
-      const { isHovering } = state
-
-      const mergedClasses = cx('ui-list__item__content', classes.content)
-
-      return !content && !contentMedia ? null : (
-        <Layout
-          className={mergedClasses}
-          alignItems="start"
-          gap={pxToRem(8)}
-          debug={debug}
-          // disappearing={!truncateContent}
-          truncateMain={truncateContent}
-          rootCSS={selection && isHovering ? { color: 'inherit' } : undefined}
-          main={content}
-          end={!isHovering && contentMedia}
-        />
-      )
-    },
   }
 
   state: any = {}
@@ -168,27 +83,60 @@ class ListItem extends UIComponent<any, any> {
     this.setState({ isHovering: false })
   }
 
-  renderComponent({ ElementType, classes, accessibility, rest }) {
-    const { as, debug, endMedia, media, renderMainArea } = this.props
-    const { isHovering } = this.state
+  renderComponent({ ElementType, classes, accessibility, rest, styles }) {
+    const {
+      as,
+      debug,
+      endMedia,
+      media,
+      content,
+      contentMedia,
+      header,
+      headerMedia,
+      selection,
+      truncateContent,
+      truncateHeader,
+    } = this.props
 
-    const startArea = media
-    const mainArea = renderMainArea(this.props, this.state, classes)
+    const { isHovering } = this.state
     const endArea = isHovering && endMedia
 
+    const hoveringSelectionCSS = selection && isHovering ? { color: 'inherit' } : {}
+
+    const headerCSS = {
+      ...styles.header,
+      ...hoveringSelectionCSS,
+    }
+    const headerMediaCSS = {
+      ...styles.headerMedia,
+      ...hoveringSelectionCSS,
+    }
+    const contentCSS = {
+      ...styles.content,
+      ...hoveringSelectionCSS,
+    }
+
     return (
-      <Layout
+      <ItemLayout
         as={as}
-        alignItems="center"
-        gap={pxToRem(8)}
         className={classes.root}
+        rootCSS={styles.root}
+        content={content}
+        contentMedia={!isHovering && contentMedia}
         debug={debug}
-        reducing
-        start={startArea}
-        main={mainArea}
-        end={endArea}
+        endMedia={endArea}
+        header={header}
+        headerMedia={headerMedia}
+        media={media}
+        mediaCSS={styles.media}
+        selection={selection}
+        truncateContent={truncateContent}
+        truncateHeader={truncateHeader}
         onMouseEnter={this.handleMouseEnter}
         onMouseLeave={this.handleMouseLeave}
+        headerCSS={headerCSS}
+        headerMediaCSS={headerMediaCSS}
+        contentCSS={contentCSS}
         {...accessibility.attributes.root}
         {...rest}
       />