+++ src/components/Provider/ProviderConsumer.tsx
@@ -2,20 +2,16 @@ import * as PropTypes from 'prop-types'
 import * as React from 'react'
 import { FelaTheme } from 'react-fela'
 
-// TODO: TypeScript incorrectly compiles typings defined in component files
-// TODO: see: https://github.com/stardust-ui/react/issues/87#issuecomment-412657622
-//
-// import { IThemePrepared } from '../../../types/theme'
-//
-// export interface IProviderConsumerProps {
-//   render: (theme: IThemePrepared) => React.ReactNode
-// }
+import { IThemePrepared } from '../../../types/theme'
+
+export interface IProviderConsumerProps {
+  render: (theme: IThemePrepared) => React.ReactNode
+}
 
 /**
  * The Provider's Consumer is for accessing the theme.
  */
-// TODO: restore provider consumer interface once above bug is fixed
-const ProviderConsumer: React.SFC<any> = props => <FelaTheme {...props} />
+const ProviderConsumer: React.SFC<IProviderConsumerProps> = props => <FelaTheme {...props} />
 
 ProviderConsumer.propTypes = {
   /**