+++ CHANGELOG.md
@@ -17,8 +17,19 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### BREAKING CHANGES
+- Remove Children API support from Menu component @miroslavstastny ([#111](https://github.com/stardust-ui/react/pull/111))
+
+### Fixes
+- Fix wrong typings generated for dist @kuzhelov ([#99](https://github.com/stardust-ui/react/pull/99))
+- Fix components generation script @kuzhelov ([#105](https://github.com/stardust-ui/react/pull/105))
+- Reactivate tests for `Text` @kuzhelov ([#104](https://github.com/stardust-ui/react/pull/104))
+- Fix Button icon color @levithomason ([#102](https://github.com/stardust-ui/react/pull/102))
+
 ### Features
-- Introduce basic `Segment` component @kuzhelov ([#103](https://github.com/stardust-ui/react/pull/103))
+- Add Menu `iconOnly`, MenuItem `iconOnly` and `icon` props @miroslavstastny ([#73](https://github.com/stardust-ui/react/pull/73))
+- Add `Grid` component base implementation @Bugaa92 ([#93](https://github.com/stardust-ui/react/pull/93))
+- Add basic `Segment` component @kuzhelov ([#103](https://github.com/stardust-ui/react/pull/103))
 
 <!--------------------------------[ v0.2.7 ]------------------------------- -->
 ## [v0.2.7](https://github.com/stardust-ui/react/tree/v0.2.7) (2018-08-13)
@@ -35,6 +46,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Features
 - Add `color` variables to Header and Header.Description @kuzhelov ([#72](https://github.com/stardust-ui/react/pull/72))
+- Add `ItemLayout` component @mnajdova ([#60](https://github.com/stardust-ui/react/pull/60))
 
 <!--------------------------------[ v0.2.6 ]------------------------------- -->
 ## [v0.2.6](https://github.com/stardust-ui/react/tree/v0.2.6) (2018-08-09)